3.3. Вынесет логотип и svg в отдельный компонент
