+++ project/src/components/logo/logo.tsx
@@ -1,0 +1,18 @@
+import { Link } from 'react-router-dom';
+
+type LogoProps = {
+  lightMode: boolean;
+}
+
+function Logo({lightMode}:LogoProps): JSX.Element {
+  return (
+    <div className="logo">
+      <Link className={lightMode ? 'logo__link logo__link--light' : 'logo__link'} to="/">
+        <span className="logo__letter logo__letter--1">W</span>
+        <span className="logo__letter logo__letter--2">T</span>
+        <span className="logo__letter logo__letter--3">W</span>
+      </Link>
+    </div>
+  );
+}
+export default Logo;
